movies module mvc done
