+++ view/movies_dashboard.css
@@ -19,6 +19,33 @@ body {
 header {
     display: flex;
     background-color: var(--secondary-color);
+    justify-content: space-between;
+    align-items: center;
+}
+
+header nav {
+    align-items: center;
+    transition: 0.5s;
+}
+
+header nav ul {
+    display: inline-block;
+    font-size: 16px;
+}
+
+header nav ul li i {
+    color: white;
+    transition: all 300ms ease-in-out;
+    font-size: 25px;
+}
+
+header nav ul li i:hover {
+    color: red;
+}
+
+header nav ul li {
+    margin-right: 20px;
+    display: inline-block;
 }
 
 .search {
@@ -114,15 +141,15 @@ main {
 }
 
 .subscribe i {
-    font-size: 20px;
+    font-size: 25px;
     margin-right: 20px;
     cursor: pointer;
+    color: white;
 }
 
 .flex {
     display: flex;
 }
-
 .subscribe button {
     background: transparent;
     border: none;
@@ -139,10 +166,13 @@ main {
 
 #playbtn {
     background: #f2b704;
-    color: white;
     width: 40px;
     height: 40px;
     line-height: 40px;
     text-align: center;
     border-radius: 50%;
+}
+
+#refreshbtn {
+    font-size: 25px;
 }
